feat: change Quantity from int to decimal for fractional units
Allows products and stock movements to be tracked in non-integer
quantities — required for by-weight items (bread by gram, deli by kg,
bulk goods, fractional portions). Existing integer quantities
remain valid since int is implicitly convertible to decimal.

Affected:
- Product (PiecesPerBatch / stock fields), SaleDetail (Quantity),
  StockMovement (Quantity)
- All matching DTOs (Product, SaleDetail, SaleItem, SpecialProductItem,
  StockAddition, StockMovement, TopSellingProduct)
- EF migration 20251229182628_ChangeQuantityToDecimal alters the
  underlying SQL column types

+++ BakeryPOS.API/BakeryPOS.API/DTOs/SpecialProductItemDto.cs
@@ -3,8 +3,8 @@
     public class SpecialProductItemDto
     {
         public string ProductName { get; set; } = string.Empty;
-        public int QuantityAdded { get; set; }
-        public int QuantitySold { get; set; }
+        public decimal QuantityAdded { get; set; }
+        public decimal QuantitySold { get; set; }
         public decimal TotalRevenue { get; set; }
         public decimal TotalCost { get; set; }
         public decimal Profit { get; set; }
