+++ BakeryPOS.API/BakeryPOS.API/DTOs/StockMovementDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public DateTime Timestamp { get; set; }
-        public int QuantityChange { get; set; }
+        public decimal QuantityChange { get; set; }
         public string Type { get; set; } = string.Empty; // e.g., "Addition", "Sale"
         public string ProductName { get; set; } = string.Empty;
         public string UserName { get; set; } = string.Empty;
